Progressbar stylesheet cleanup
