+++ src/cfclient/ui/main.py
@@ -39,6 +39,7 @@ from cfclient.utils.config import Config
 from cfclient.utils.config_manager import ConfigManager
 from cfclient.utils.input import JoystickReader
 from cfclient.utils.logconfigreader import LogConfigReader
+from cfclient.utils.ui import UiUtils
 from cfclient.utils.zmq_led_driver import ZMQLEDDriver
 from cfclient.utils.zmq_param import ZMQParamAccess
 from cflib.crazyflie import Crazyflie
@@ -95,24 +96,6 @@ class BatteryStates:
     BATTERY, CHARGING, CHARGED, LOW_POWER = list(range(4))
 
 
-COLOR_BLUE = '#3399ff'
-COLOR_GREEN = '#00ff60'
-COLOR_RED = '#cc0404'
-
-
-def progressbar_stylesheet(color):
-    return """
-        QProgressBar {
-            border: 1px solid #333;
-            background-color: transparent;
-        }
-
-        QProgressBar::chunk {
-            background-color: """ + color + """;
-        }
-    """
-
-
 class MainUI(QtWidgets.QMainWindow, main_window_class):
     connectionLostSignal = pyqtSignal(str, str)
     connectionInitiatedSignal = pyqtSignal(str)
@@ -136,42 +119,6 @@ class MainUI(QtWidgets.QMainWindow, main_window_class):
             self.resize(size[0], size[1])
         except KeyError:
             pass
-
-        ######################################################
-        # By lxrocks
-        # 'Skinny Progress Bar' tweak for Yosemite
-        # Tweak progress bar - artistic I am not - so pick your own colors !!!
-        # Only apply to Yosemite
-        ######################################################
-        import platform
-
-        if platform.system() == 'Darwin':
-
-            (Version, junk, machine) = platform.mac_ver()
-            logger.info("This is a MAC - checking if we can apply Progress "
-                        "Bar Stylesheet for Yosemite Skinny Bars ")
-            yosemite = (10, 10, 0)
-            tVersion = tuple(map(int, (Version.split("."))))
-
-            if tVersion >= yosemite:
-                logger.info("Found Yosemite - applying stylesheet")
-
-                tcss = """
-                    QProgressBar {
-                        border: 1px solid grey;
-                        border-radius: 5px;
-                        text-align: center;
-                    }
-                    QProgressBar::chunk {
-                        background-color: """ + COLOR_BLUE + """;
-                    }
-                 """
-                self.setStyleSheet(tcss)
-
-            else:
-                logger.info("Pre-Yosemite - skinny bar stylesheet not applied")
-
-        ######################################################
 
         self.cf = Crazyflie(ro_cache=None,
                             rw_cache=cfclient.config_path + "/cache")
@@ -275,10 +222,7 @@ class MainUI(QtWidgets.QMainWindow, main_window_class):
         self._log_error_signal.connect(self._logging_error)
 
         self.batteryBar.setTextVisible(False)
-        self.batteryBar.setStyleSheet(progressbar_stylesheet(COLOR_BLUE))
-
         self.linkQualityBar.setTextVisible(False)
-        self.linkQualityBar.setStyleSheet(progressbar_stylesheet(COLOR_BLUE))
 
         # Connect link quality feedback
         self.cf.link_quality_updated.add_callback(self.linkQualitySignal.emit)
@@ -559,15 +503,15 @@ class MainUI(QtWidgets.QMainWindow, main_window_class):
     def _update_battery(self, timestamp, data, logconf):
         self.batteryBar.setValue(int(data["pm.vbat"] * 1000))
 
-        color = COLOR_BLUE
+        color = UiUtils.COLOR_BLUE
         # TODO firmware reports fully-charged state as 'Battery',
         # rather than 'Charged'
         if data["pm.state"] in [BatteryStates.CHARGING, BatteryStates.CHARGED]:
-            color = COLOR_GREEN
+            color = UiUtils.COLOR_GREEN
         elif data["pm.state"] == BatteryStates.LOW_POWER:
-            color = COLOR_RED
+            color = UiUtils.COLOR_RED
 
-        self.batteryBar.setStyleSheet(progressbar_stylesheet(color))
+        self.batteryBar.setStyleSheet(UiUtils.progressbar_stylesheet(color))
         self._aff_volts.setText(("%.3f" % data["pm.vbat"]))
 
     def _connected(self):
